requirements.txt
ipython==7.34.0
keras==2.9.0
matplotlib==3.7.1
numpy==1.24.1
opencv_python==4.6.0.66
Pillow==9.5.0
scikit_learn==1.2.2
tensorflow==2.9.1
Flask==2.3.3
flask-cors==6.0.1